+++ src/block-kit/Blocks.tsx
@@ -1,8 +1,12 @@
 /** @jsx JSXSlack.h */
-import { JSXSlack, jsxOnParsed } from '../jsx'
-import { ArrayOutput, aliasTo, isNode, wrap } from '../utils'
+import { JSXSlack } from '../jsx'
+import { ArrayOutput, aliasTo, isNode } from '../utils'
+import { actionTypes } from './Actions'
 import { internalHiddenType, internalSubmitType } from './Input'
+import { sectionAccessoryTypes } from './Section'
 import { Divider, Image, Input, Section, Select, Textarea } from './index'
+
+type KnownMap = Map<InternalBlockType | undefined, (string | symbol)[]>
 
 export interface BlocksProps {
   children: JSXSlack.Children<BlockComponentProps>
@@ -20,15 +24,19 @@ export enum InternalBlockType {
 
 export const blockTypeSymbol = Symbol('jsx-slack-block-type')
 
-const knownBlocks: Map<
-  InternalBlockType | undefined,
-  (string | symbol)[]
-> = new Map()
+const knownBlocks: KnownMap = new Map()
+const knownActions: KnownMap = new Map()
+const knownSectionAccessories: KnownMap = new Map()
 
 const basicBlocks = ['actions', 'context', 'divider', 'image', 'section']
 
 // Blocks
 knownBlocks.set(undefined, [...basicBlocks, 'file'])
+knownActions.set(undefined, actionTypes.filter(t => t !== 'radio_buttons'))
+knownSectionAccessories.set(
+  undefined,
+  sectionAccessoryTypes.filter(t => t !== 'radio_buttons')
+)
 
 // Modal
 knownBlocks.set(InternalBlockType.Modal, [
@@ -37,15 +45,30 @@ knownBlocks.set(InternalBlockType.Modal, [
   internalHiddenType,
   internalSubmitType,
 ])
+knownActions.set(
+  InternalBlockType.Modal,
+  actionTypes.filter(t => t !== 'radio_buttons')
+)
+knownSectionAccessories.set(
+  InternalBlockType.Modal,
+  sectionAccessoryTypes.filter(t => t !== 'radio_buttons')
+)
 
 // Home
 knownBlocks.set(InternalBlockType.Home, basicBlocks)
+knownActions.set(InternalBlockType.Home, [...actionTypes])
+knownSectionAccessories.set(InternalBlockType.Home, [...sectionAccessoryTypes])
 
 export const Blocks: JSXSlack.FC<BlocksProps> = props => {
   const internalType: InternalBlockType | undefined = props[blockTypeSymbol]
 
-  const normalized = wrap(props.children).map(child => {
-    if (child && isNode(child) && typeof child.type === 'string') {
+  const availableBlocks = knownBlocks.get(internalType) || []
+  const availableActions = knownActions.get(internalType) || []
+  const availableSectionAccessories =
+    knownSectionAccessories.get(internalType) || []
+
+  const normalized = JSXSlack.normalizeChildren(props.children).map(child => {
+    if (isNode(child) && typeof child.type === 'string') {
       // Aliasing intrinsic elements to Block component
       switch (child.type) {
         case 'hr':
@@ -69,19 +92,40 @@ export const Blocks: JSXSlack.FC<BlocksProps> = props => {
     return child
   })
 
-  const node = <ArrayOutput>{normalized}</ArrayOutput>
+  for (const child of normalized) {
+    let valid = false
 
-  node.props[jsxOnParsed] = parsed => {
-    // Check the final output
-    const availableBlocks = knownBlocks.get(internalType) || []
+    if (isNode(child) && child.type === JSXSlack.NodeType.object) {
+      if (availableBlocks.includes(child.props.type)) {
+        // Validate compatibillity of actions and section accessory with container
+        if (
+          child.props.type === 'actions' &&
+          child.props.elements.some(e => !availableActions.includes(e.type))
+        )
+          throw new Error(
+            '<Actions> block has an incompatible element with container.'
+          )
 
-    if (parsed.some(b => !availableBlocks.includes(b.type)))
+        if (
+          child.props.type === 'section' &&
+          child.props.accessory &&
+          !availableSectionAccessories.includes(child.props.accessory.type)
+        )
+          throw new Error(
+            '<Section> block has an incompatible element with container.'
+          )
+
+        valid = true
+      }
+    }
+
+    if (!valid)
       throw new Error(
         'Block container allows only including valid layout block components.'
       )
   }
 
-  return node
+  return <ArrayOutput>{normalized}</ArrayOutput>
 }
 
 export const BlocksInternal: JSXSlack.FC<