+++ test/block-kit/block-elements/interactive-components.tsx
@@ -14,10 +14,14 @@ import JSXSlack, {
   ConversationsSelect,
   DatePicker,
   ExternalSelect,
+  Home,
+  Modal,
   Optgroup,
   Option,
   Overflow,
   OverflowItem,
+  RadioButton,
+  RadioButtonGroup,
   Section,
   Select,
   UsersSelect,
@@ -557,7 +561,7 @@ describe('Interactive components', () => {
         )
       ).toStrictEqual([action(baseOverflow)])
 
-      // confirm prop and HTML-compatible props
+      // "confirm" prop and HTML-compatible props
       expect(
         JSXSlack(
           <Blocks>
@@ -641,6 +645,171 @@ describe('Interactive components', () => {
           </Blocks>
         )
       ).toStrictEqual([datePickerAction])
+    })
+  })
+
+  describe('<RadioButtonGroup>', () => {
+    it('outputs radio button group in actions block', () => {
+      const radioButtonAction = {
+        type: 'radio_buttons',
+        action_id: 'radio-buttons',
+        options: [
+          {
+            text: { type: 'plain_text', text: '1st', emoji: true },
+            description: {
+              type: 'plain_text',
+              text: 'The first option',
+              emoji: true,
+            },
+            value: 'first',
+          },
+          {
+            text: { type: 'plain_text', text: '2nd', emoji: true },
+            description: {
+              type: 'plain_text',
+              text: 'The second option',
+              emoji: true,
+            },
+            value: 'second',
+          },
+          {
+            text: { type: 'plain_text', text: '3rd', emoji: true },
+            value: 'third',
+          },
+        ],
+        initial_option: {
+          text: { type: 'plain_text', text: '2nd', emoji: true },
+          description: {
+            type: 'plain_text',
+            text: 'The second option',
+            emoji: true,
+          },
+          value: 'second',
+        },
+      }
+
+      expect(
+        JSXSlack(
+          <Home>
+            <Actions blockId="actions">
+              <RadioButtonGroup actionId="radio-buttons" value="second">
+                <RadioButton value="first" description="The first option">
+                  1st
+                </RadioButton>
+                <RadioButton value="second" description="The second option">
+                  2nd
+                </RadioButton>
+                <RadioButton value="third">3rd</RadioButton>
+              </RadioButtonGroup>
+            </Actions>
+          </Home>
+        ).blocks
+      ).toStrictEqual([action(radioButtonAction)])
+
+      // "confirm" prop and HTML-compatible props
+      expect(
+        JSXSlack(
+          <Home>
+            <Actions id="actions">
+              <RadioButtonGroup
+                name="radio-buttons"
+                value="second"
+                confirm={
+                  <Confirm title="a" confirm="b" deny="c">
+                    foobar
+                  </Confirm>
+                }
+              >
+                <RadioButton value="first" description="The first option">
+                  1st
+                </RadioButton>
+                <RadioButton value="second" description="The second option">
+                  2nd
+                </RadioButton>
+                <RadioButton value="third">3rd</RadioButton>
+              </RadioButtonGroup>
+            </Actions>
+          </Home>
+        ).blocks
+      ).toStrictEqual([
+        action({
+          ...radioButtonAction,
+          confirm: {
+            title: { type: 'plain_text', text: 'a', emoji: true },
+            confirm: { type: 'plain_text', text: 'b', emoji: true },
+            deny: { type: 'plain_text', text: 'c', emoji: true },
+            text: { type: 'mrkdwn', text: 'foobar', verbatim: true },
+          },
+        } as any),
+      ])
+    })
+
+    it('outputs radio button group in section block', () => {
+      const [section]: SectionBlock[] = JSXSlack(
+        <Home>
+          <Section>
+            test
+            <RadioButtonGroup>
+              <RadioButton value="a">A</RadioButton>
+            </RadioButtonGroup>
+          </Section>
+        </Home>
+      ).blocks
+
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+      expect(section.accessory!.type).toBe('radio_buttons')
+    })
+
+    it('throws error when <RadioButtonGroup> has not contained <RadioButton>', () => {
+      expect(() =>
+        JSXSlack(
+          <Home>
+            <Actions>
+              <RadioButtonGroup>{}</RadioButtonGroup>
+            </Actions>
+          </Home>
+        )
+      ).toThrow(/must include/i)
+
+      expect(() =>
+        JSXSlack(
+          <Home>
+            <Actions>
+              <RadioButtonGroup>
+                <Option value="wtf">I'm not radio button</Option>
+              </RadioButtonGroup>
+            </Actions>
+          </Home>
+        )
+      ).toThrow(/must include/i)
+    })
+
+    it('throws error when using <RadioButtonGroup> within invalid container', () => {
+      const invalidContainers: JSXSlack.FC<{ children: any }>[] = [
+        Blocks,
+        ({ children }) => <Modal title="test">{children}</Modal>,
+      ]
+
+      const invalidBlocks: JSXSlack.FC<{ children: any }>[] = [
+        Actions,
+        ({ children }) => <Section>test{children}</Section>,
+      ]
+
+      for (const Container of invalidContainers) {
+        for (const Block of invalidBlocks) {
+          expect(() =>
+            JSXSlack(
+              <Container>
+                <Block>
+                  <RadioButtonGroup>
+                    <RadioButton value="a">A</RadioButton>
+                  </RadioButtonGroup>
+                </Block>
+              </Container>
+            )
+          ).toThrow(/incompatible/i)
+        }
+      }
     })
   })
 })